add /blog/comments page for viewing all blog comments
