+++ app/core/Controller/BlogController.php
@@ -3,14 +3,14 @@
 namespace Blog\Controller;
 
 use Blog\Modules\Entity\BlogArticle;
-use Blog\Modules\Entity\Comment;
 use Blog\Modules\TemplateFacade\Form;
 use Blog\Modules\View\Blog;
 use Blog\Request\BlogArticleCreateRequest;
-use Blog\Request\CommentRequest;
 
 class BlogController extends BaseController
 {
+    use Components\BlogControllerComments;
+
     protected \Blog\Modules\Template\Element $content;
 
     protected int $status;
@@ -47,8 +47,9 @@ class BlogController extends BaseController
             }
             return Blog::viewBlogArticle($argument);
         }
+        // if no arguments passed to blog controller then render blog articles list
         // set page meta
-        app()->page()->setMetaTitle(strToken('Журнал веб-разработчика | :[site]'));
+        app()->page()->setMetaTitle(stok('Журнал веб-разработчика | :[site]'));
         app()->page()->setMeta('description', [
             'name' => 'description',
             'content' => 'Полезные и интересные материалы и статьи в персональном блоге веб-разработчика'
@@ -57,9 +58,13 @@ class BlogController extends BaseController
         return true;
     }
 
-    public function getTitle(): string
+    public function getTitle(): ControlledPageTitle
     {
-        return t('blog');
+        parent::getTitle();
+        if (!$this->title->isset()) {
+            $this->title->set(t('blog'));
+        }
+        return $this->title;
     }
 
     public function postRequest(): void
@@ -75,49 +80,6 @@ class BlogController extends BaseController
         exit;
     }
 
-    protected function postRequestCommentAdd(array $data): void
-    {
-        $request = new CommentRequest($data);
-        if ($request->isValid()) {
-            $result = Comment::create($request);
-        } else {
-            $result = null;
-        }
-        if ($result) {
-            msgr()->notice('Ваш комментарий отправлен и будет рассмотрен в ближайшее время.');
-        } else if (!is_null($result)) {
-            msgr()->error('При отправке комментария возникла ошибка. Если проблема повторяется, пожалуйста @contact_me.', ['contact_me' => '<a href="' . tpllink('contacts') . '">свяжитесь со мной</a>']);
-        }
-        $redirect = ($data['article_id'] ?? false) ? '/blog/' . $data['article_id'] : '<current>';
-        app()->router()->redirect($redirect);
-        return;
-    }
-
-    protected function getRequestComment(): bool
-    {
-        // verify user access level
-        if (!app()->user()->verifyAccessLevel(4)) {
-            $this->status = 403;
-            return false;
-        }
-        // check request arguments
-        $cid = app()->router()->arg(3);
-        $action = app()->router()->arg(4);
-        if (!$cid || !$action) {
-            $this->status = 404;
-            return false;
-        }
-        $action = pascalCase($action);
-        $comment = new Comment($cid);
-        if (!$comment->exists() || !method_exists($comment, $action)) {
-            $this->status = 404;
-            return false;
-        }
-        $comment->$action();
-        app()->router()->redirect('<previous>');
-        return true;
-    }
-
     protected function getRequestCreate(): bool
     {
         // verify user access level
@@ -126,7 +88,8 @@ class BlogController extends BaseController
             return false;
         }
         $form = new Form('blog-article');
-        app()->page()->setTitle('Создание нового материала для блога');
+        // app()->page()->setTitle('Создание нового материала для блога');
+        $this->getTitle()->set('Создание нового материала для блога');
         app()->page()->addContent($form);
         return true;
     }
